This is still work in progress. So don't use it on production systems. These changes also are not yet put back into the FreeBSD base code. But I need a backup...

+++ trunk/KERN/nke/sctp_nke/SnowLeopard/netinet/sctp_bsd_addr.c
@@ -651,7 +651,7 @@ out:
 #endif
 
 struct mbuf *
-sctp_get_mbuf_for_msg(unsigned int space_needed, int want_header,
+sctp_get_mbuf_for_msg(size_t space_needed, int want_header,
 		      int how, int allonebuf, int type)
 {
     struct mbuf *m = NULL;
