+++ trunk/KERN/nke/sctp_nke/SnowLeopard/netinet/sctputil.c
@@ -6657,7 +6657,7 @@ sctp_connectx_helper_add(struct sctp_tcb *stcb, struct sockaddr *addr,
 struct sctp_tcb *
 sctp_connectx_helper_find(struct sctp_inpcb *inp, struct sockaddr *addr,
 			  int *totaddr, int *num_v4, int *num_v6, int *error,
-			  int limit, int *bad_addr)
+			  size_t limit, int *bad_addr)
 {
 	struct sockaddr *sa;
 	struct sctp_tcb *stcb=NULL;
